// src/YGOCore/types/commands.ts
export type CommandType = "NULL"
    | "Normal Summon"
    | "Tribute Summon"
    | "Tribute Set"
    | "Special Summon"
    | "Banish"
    | "Banish FD"
    | "Shuffle Deck"
    | "Start Hand"
    | "Draw From Deck"
    | "Mill From Deck"
    | "Fusion Summon"
    | "Synchro Summon"
    | "Link Summon"
    | "XYZ Summon"
    | "XYZ Attach Material"
    | "XYZ Detach Material"
    | "Set Monster"
    | "Set ST"
    | "To ST"
    | "To Extra Deck"
    | "Send To GY"
    | "Destroy"
    | "Activate"
    | "Change Battle Position"
    | "Move Card"
    | "To Top Deck"
    | "To Bottom Deck"
    | "To Hand"
    | "Reveal"
    | "Target"
    | "Field Spell"
    | "Change Card Position"
    | "Change Card Atk Def"
    | "Flip";

export interface Command {
    type: CommandType
    parent: Command | null;
    commandId: number;
    init(ygo: YGOCore): void
    isValid(): boolean
    exec(): void
    undo(): void
    toJSON<T extends any>(): { type: string, data: T }
}
// ...
export interface XYZSummonCommandData {
    player: number
    id: number
    originZone: FieldZone
    zone: FieldZone
    position?: CardPosition
    materials: Array<{
        id: number
        zone: FieldZone
    }>
}